deployment/requirement.txt
numpy
opencv-python
pillow
flask
requests
paho-mqtt
tflite-runtime
